# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  controller.CommonResult:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
  controller.Page:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  controller.ResultWithPage:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
      page:
        $ref: '#/definitions/controller.Page'
    type: object
...
  message.ApplyParameterGroupReq:
    properties:
      clusterId:
        example: "123"
        type: string
      reboot:
        type: boolean
    required:
    - clusterId
    type: object
  message.ApplyParameterGroupResp:
    properties:
      clusterId:
        example: "123"
        type: string
      paramGroupId:
        example: "123"
        type: string
      workFlowId:
        description: Asynchronous task workflow ID
        type: string
    type: object
...
  message.CopyParameterGroupReq:
    properties:
      name:
        example: 8C16GV4_copy
        type: string
      note:
        example: copy param group
        type: string
    required:
    - name
    type: object
  message.CopyParameterGroupResp:
    properties:
      paramGroupId:
        example: "1"
        type: string
    type: object
...
  message.CreateParameterGroupReq:
    properties:
      addParams:
        items:
          $ref: '#/definitions/message.ParameterInfo'
        type: array
      clusterSpec:
        example: 8C16G
        type: string
      clusterVersion:
        example: v5.0
        type: string
      dbType:
        enum:
        - 1
        - 2
        example: 1
        type: integer
      groupType:
        enum:
        - 1
        - 2
        example: 1
        type: integer
      name:
        example: 8C16GV4_default
        type: string
      note:
        example: default param group
        type: string
      params:
        items:
          $ref: '#/definitions/structs.ParameterGroupParameterSampleInfo'
        type: array
    required:
    - clusterSpec
    - clusterVersion
    - dbType
    - groupType
    - name
    - params
    type: object
  message.CreateParameterGroupResp:
    properties:
      paramGroupId:
        example: "1"
        type: string
    type: object
...
  message.DeleteParameterGroupResp:
    properties:
      paramGroupId:
        example: "1"
        type: string
    type: object
...
  message.DetailParameterGroupResp:
    properties:
      clusterSpec:
        example: 8C16G
        type: string
      clusterVersion:
        example: v5.0
        type: string
      createTime:
        example: 1636698675
        type: integer
      dbType:
        enum:
        - 1
        - 2
        example: 1
        type: integer
      groupType:
        enum:
        - 1
        - 2
        example: 0
        type: integer
      hasDefault:
        enum:
        - 1
        - 2
        example: 1
        type: integer
      name:
        example: default
        type: string
      note:
        example: default param group
        type: string
      paramGroupId:
        example: "1"
        type: string
      params:
        items:
          $ref: '#/definitions/structs.ParameterGroupParameterInfo'
        type: array
      updateTime:
        example: 1636698675
        type: integer
    type: object
...
  message.ParameterInfo:
    properties:
      category:
        example: log
        type: string
      defaultValue:
        example: "1024"
        type: string
      description:
        example: binlog size
        type: string
      hasApply:
        enum:
        - 0
        - 1
        example: 1
        type: integer
      hasReboot:
        enum:
        - 0
        - 1
        example: 0
        type: integer
      instanceType:
        example: TiDB
        type: string
      name:
        example: binlog_size
        type: string
      note:
        type: string
      range:
        items:
          type: string
        type: array
      rangeType:
        enum:
        - 0
        - 1
        - 2
        example: 1
        type: integer
      readOnly:
        enum:
        - 0
        - 1
        example: 0
        type: integer
      systemVariable:
        example: log.binlog_size
        type: string
      type:
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        example: 0
        type: integer
      unit:
        example: MB
        type: string
      unitOptions:
        example:
        - KB
        - MB
        - GB
        items:
          type: string
        type: array
      updateSource:
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        example: 0
        type: integer
    type: object
...
  message.QueryParameterGroupResp:
    properties:
      clusterSpec:
        example: 8C16G
        type: string
      clusterVersion:
        example: v5.0
        type: string
      createTime:
        example: 1636698675
        type: integer
      dbType:
        enum:
        - 1
        - 2
        example: 1
        type: integer
      groupType:
        enum:
        - 1
        - 2
        example: 0
        type: integer
      hasDefault:
        enum:
        - 1
        - 2
        example: 1
        type: integer
      name:
        example: default
        type: string
      note:
        example: default param group
        type: string
      paramGroupId:
        example: "1"
        type: string
      params:
        items:
          $ref: '#/definitions/structs.ParameterGroupParameterInfo'
        type: array
      updateTime:
        example: 1636698675
        type: integer
    type: object
...
  message.UpdateParameterGroupReq:
    properties:
      addParams:
        items:
          $ref: '#/definitions/message.ParameterInfo'
        type: array
      clusterSpec:
        example: 8C16G
        type: string
      clusterVersion:
        example: v5.0
        type: string
      delParams:
        example:
        - "1"
        items:
          type: string
        type: array
      name:
        example: 8C16GV4_new
        type: string
      note:
        example: update param group
        type: string
      params:
        items:
          $ref: '#/definitions/structs.ParameterGroupParameterSampleInfo'
        type: array
    required:
    - params
    type: object
  message.UpdateParameterGroupResp:
    properties:
      paramGroupId:
        example: "1"
        type: string
    type: object
...
  structs.ParameterGroupParameterInfo:
    properties:
      category:
        example: basic
        type: string
      createTime:
        example: 1636698675
        type: integer
      defaultValue:
        example: "1"
        type: string
      description:
        example: binlog cache size
        type: string
      hasApply:
        enum:
        - 0
        - 1
        example: 1
        type: integer
      hasReboot:
        enum:
        - 0
        - 1
        example: 0
        type: integer
      instanceType:
        example: tidb
        type: string
      name:
        example: binlog_size
        type: string
      note:
        example: binlog cache size
        type: string
      paramId:
        example: "1"
        type: string
      range:
        example:
        - "1"
        - ' 1000'
        items:
          type: string
        type: array
      rangeType:
        enum:
        - 0
        - 1
        - 2
        example: 1
        type: integer
      readOnly:
        enum:
        - 0
        - 1
        example: 0
        type: integer
      systemVariable:
        example: log.log_level
        type: string
      type:
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        example: 0
        type: integer
      unit:
        example: MB
        type: string
      unitOptions:
        example:
        - KB
        - MB
        - GB
        items:
          type: string
        type: array
      updateSource:
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        example: 0
        type: integer
      updateTime:
        example: 1636698675
        type: integer
    type: object
  structs.ParameterGroupParameterSampleInfo:
    properties:
      defaultValue:
        example: "1"
        type: string
      note:
        example: binlog cache size
        type: string
      paramId:
        example: "123"
        type: string
    required:
    - defaultValue
    - paramId
    type: object
...
host: localhost:4116
info:
  contact:
    email: zhangpeijin@pingcap.com
    name: zhangpeijin
  description: EM UI API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: EM UI API
  version: "1.0"
paths:
...
  /param-groups/:
    get:
      consumes:
      - application/json
      description: query parameter group
      parameters:
      - example: 8C16G
        in: query
        name: clusterSpec
        type: string
      - example: v5.0
        in: query
        name: clusterVersion
        type: string
      - enum:
        - 0
        - 1
        - 2
        example: 0
        in: query
        name: dbType
        type: integer
      - enum:
        - 0
        - 1
        - 2
        example: 0
        in: query
        name: hasDefault
        type: integer
      - example: false
        in: query
        name: hasDetail
        type: boolean
      - example: default
        in: query
        name: name
        type: string
      - description: Current page location
        in: query
        name: page
        type: integer
      - description: Number of this request
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.ResultWithPage'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/message.QueryParameterGroupResp'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: query parameter group
      tags:
      - parameter group
    post:
      consumes:
      - application/json
      description: create a parameter group
      parameters:
      - description: create request
        in: body
        name: createReq
        required: true
        schema:
          $ref: '#/definitions/message.CreateParameterGroupReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.CreateParameterGroupResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: create a parameter group
      tags:
      - parameter group
  /param-groups/{paramGroupId}:
    delete:
      consumes:
      - application/json
      description: delete a parameter group
      parameters:
      - description: parameter group id
        in: path
        name: paramGroupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.DeleteParameterGroupResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: delete a parameter group
      tags:
      - parameter group
    get:
      consumes:
      - application/json
      description: show details of a parameter group
      parameters:
      - description: parameter group id
        in: path
        name: paramGroupId
        required: true
        type: string
      - in: query
        name: instanceType
        type: string
      - in: query
        name: paramName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.DetailParameterGroupResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: show details of a parameter group
      tags:
      - parameter group
    put:
      consumes:
      - application/json
      description: update a parameter group
      parameters:
      - description: parameter group id
        in: path
        name: paramGroupId
        required: true
        type: string
      - description: update parameter group request
        in: body
        name: updateReq
        required: true
        schema:
          $ref: '#/definitions/message.UpdateParameterGroupReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.UpdateParameterGroupResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: update a parameter group
      tags:
      - parameter group
  /param-groups/{paramGroupId}/apply:
    post:
      consumes:
      - application/json
      description: apply a parameter group
      parameters:
      - description: parameter group id
        in: path
        name: paramGroupId
        required: true
        type: string
      - description: apply parameter group request
        in: body
        name: applyReq
        required: true
        schema:
          $ref: '#/definitions/message.ApplyParameterGroupReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.ApplyParameterGroupResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: apply a parameter group
      tags:
      - parameter group
  /param-groups/{paramGroupId}/copy:
    post:
      consumes:
      - application/json
      description: copy a parameter group
      parameters:
      - description: parameter group id
        in: path
        name: paramGroupId
        required: true
        type: string
      - description: copy parameter group request
        in: body
        name: copyReq
        required: true
        schema:
          $ref: '#/definitions/message.CopyParameterGroupReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.CopyParameterGroupResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: copy a parameter group
      tags:
      - parameter group
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"